+++ src/constants.rs
@@ -56,6 +56,12 @@ pub static VCF_FORMAT_NUMBER: [&[u8]; 6] = [b"1", b"2", b"A", b"R", b"G", b"."];
 /// Length of vcf string
 pub static VCF_STRING_LENGTH: usize = 5;
 
+/// CIGAR SAM
+pub static CIGAR_SAM: [u8; 9] = *b"MIDNSHP=X";
+
+/// CIGAR GFF
+pub static CIGAR_GFF: [u8; 5] = *b"MIDFR";
+
 /// biotest version
 pub const BIOTEST_VERSION: &[u8] = env!("CARGO_PKG_VERSION").as_bytes();